Handle missing input file explicitly
Check for and raise an exception when the user passes a
non-existing file as input. This makes such errors easier
to identify, as opposed to the code path and exception that
is raised with a non-existing file is passed (which also happens
to be the same exception raised when a non-XLSX file is passed,
making this error confusing).