stdin: fix false no file test and add tests for stdin
Commit 5a14efcef broke input from stdin, with the check added for
actually existing files. Add an exception when the file input is "-" so
reading from stdin works again. The original commit that introduced this
feature did not include documentation. Add documentation to "--help",
manpages and the README. Add testing for reading from stdin with python3
so future commits will not break the feature unintentionally.

Fixes: #270